fix: remove default args for non-lambda gateway
